update useEffect in Tracker
it checks if this is the first run

if useRef is true and workigis true it wont add anything

then it sets the prev working to the current worrking.

as the timer continues working moves to false.  but the prevWorking will stay true.

so once the timer hits after 25 minues tht tracker updates